# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform God API",
    "description": "REST API for agent-driven repository governance",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/agents/{agent_name}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get Agent",
        "description": "Get detailed information about a specific agent.\n\nArgs:\n    agent_name: Name of the agent to retrieve\n\nReturns:\n    AgentResponse with detailed agent information\n\nRaises:\n    NotFoundError: If agent does not exist",
        "operationId": "get_agent_api_v1_agents__agent_name__get",
        "parameters": [
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/execute": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Execute Agent",
        "description": "Execute a single agent.\n\nExecution happens synchronously for now. For long-running agents,\nconsider using the chains endpoint or implement async execution.\n\nArgs:\n    request: Agent execution request\n    background_tasks: FastAPI background tasks\n\nReturns:\n    AgentExecuteResponse with execution result\n\nRaises:\n    ValidationError: If request validation fails\n    NotFoundError: If specified agent doesn't exist",
        "operationId": "execute_agent_api_v1_agents_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentClass": {
        "type": "string",
        "enum": [
          "READ_ONLY_SCAN",
          "PLANNING_SYNTHESIS",
          "REGISTRY_STATE",
          "WRITE_GATED",
          "CONTROL_PLANE"
        ],
        "title": "AgentClass",
        "description": "Canonical agent classes defining permissions and scope."
      },
      "AgentExecuteRequest": {
        "properties": {
          "repository_root": {
            "type": "string",
            "title": "Repository Root",
            "description": "Absolute path to the repository root",
            "examples": [
              "/path/to/repository"
            ]
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "description": "Execution mode",
            "default": "dry_run"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data",
            "description": "Additional input data for the agent"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds",
            "description": "Maximum execution time in seconds",
            "default": 300
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "repository_root"
        ],
        "title": "AgentExecuteRequest",
        "description": "Request to execute a single agent."
      },
      "AgentExecuteResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/AgentExecutionResult",
            "description": "Execution result"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Execution mode used"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result",
          "mode"
        ],
        "title": "AgentExecuteResponse",
        "description": "Response model for agent execution."
      },
      "AgentExecutionResult": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name",
            "description": "Agent name"
          },
          "agent_class": {
            "$ref": "#/components/schemas/AgentClass",
            "description": "Agent class"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus",
            "description": "Execution status"
          },
          "output_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Data",
            "description": "Agent output"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Error message if failed"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Execution time in milliseconds"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Execution timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_name",
          "agent_class",
          "status",
          "timestamp"
        ],
        "title": "AgentExecutionResult",
        "description": "Result of an agent execution."
      },
...
      "AgentPermissionLevel": {
        "type": "string",
        "enum": [
          "read_only",
          "write_gated",
          "control_plane"
        ],
        "title": "AgentPermissionLevel",
        "description": "Permission levels for filesystem access."
      },
      "AgentResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Agent name"
          },
          "agent_class": {
            "$ref": "#/components/schemas/AgentClass",
            "description": "Agent class"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Agent role description"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "description": "Agent goal"
          },
          "permissions": {
            "$ref": "#/components/schemas/AgentPermissionLevel",
            "description": "Permission level"
          },
          "allowed_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Paths",
            "description": "Allowed write paths"
          },
          "disallowed_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Disallowed Paths",
            "description": "Disallowed paths"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema",
            "description": "Input schema"
          },
          "output_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Schema",
            "description": "Output schema"
          },
          "stop_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stop Conditions",
            "description": "Stop conditions"
          },
          "source_file": {
            "type": "string",
            "title": "Source File",
            "description": "Source file path"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash",
            "description": "Content hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "agent_class",
          "role",
          "goal",
          "permissions",
          "source_file",
          "content_hash"
        ],
        "title": "AgentResponse",
        "description": "Response model for a single agent."
      },
      "AgentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "stopped"
        ],
        "title": "AgentStatus",
        "description": "Possible states of agent execution."
      },
...
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "dry_run",
          "simulated",
          "live"
        ],
        "title": "ExecutionMode",
        "description": "Execution mode for agents and chains."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}